# Optimize `WorkNodeCodec`