- Do a single pass over the scheduled nodes when writing
- Don't write the node ids
- Avoid intermediate collections when writing successor node references

Combined, these changes bring the `:distributions:binZip` median time when storing from `5371.0ms` to `5155.0ms` and when loading from `1342.0ms` to `1244.5ms`.