@bamboo 2020-05-13
@bot-gradle test this